requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.30.*
jinja2==3.1.*
python-dotenv==1.0.*
httpx==0.27.*          # GitHub API calls
redis==5.0.*
rq==1.16.*
openai==1.51.*         # official SDK
semgrep==1.87.*
ruff==0.6.*
bandit==1.7.*
black==24.8.*
python-jose[cryptography]==3.3.*  # JWT handling
bcrypt>=4.0.0                      # Password hashing (direct usage, works with bcrypt 5.0+)
passlib[bcrypt]==1.7.*            # Password hashing (kept for compatibility, but using bcrypt directly)
python-multipart==0.0.*           # Form data parsing

